refactor(auth): split user actions into separate files for sign-in, sign-up, and sign-out
Refactored the contents of `user.actions.ts` by extracting the logic for each authentication action into its own dedicated file:

- `sign-in-with-credentials.action.ts`
- `sign-up-with-credentials.action.ts`
- `sign-out-user.action.ts`

This change improves modularity, readability, and maintainability of the auth-related actions. Each action is now easier to test and modify independently.
