+++ lib/actions/auth.actions/sign-out-user.action.ts
@@ -1,0 +1,7 @@
+"use server";
+
+import { signOut } from "@/auth";
+
+export async function signOutUser() {
+  await signOut();
+}
